requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.1
    # via
    #   -r requirements.in
    #   django-filter
    #   django-tenants
    #   djangorestframework
django-environ==0.10.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-tenants==3.5.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
markdown==3.4.3
    # via -r requirements.in
pillow==9.5.0
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.6.3
    # via asgiref

# The following packages are considered to be unsafe in a requirements file:
# setuptools
